Shift useChartData into unified-logs-chart-query, and refactor main UI to use that new query
