+++ apps/studio/data/logs/keys.ts
@@ -9,13 +9,23 @@ export const logsKeys = {
       'projects',
       projectRef,
       'unified-logs',
+      'logs-data',
       ...(searchParams ? [searchParams].filter(Boolean) : []),
     ] as const,
   unifiedLogsCount: (projectRef: string | undefined, searchParams: SearchParamsType | undefined) =>
     [
       'projects',
       projectRef,
-      'unified-logs-count',
+      'unified-logs',
+      'count-data',
+      ...(searchParams ? [searchParams].filter(Boolean) : []),
+    ] as const,
+  unifiedLogsChart: (projectRef: string | undefined, searchParams: SearchParamsType | undefined) =>
+    [
+      'projects',
+      projectRef,
+      'unified-logs',
+      'chart-data',
       ...(searchParams ? [searchParams].filter(Boolean) : []),
     ] as const,
 }
